fzf-tmux: Conditional install

+++ ubuntu.sh
@@ -76,6 +76,8 @@ fi
 ./ghpkg.py download
 
 # install fzf-tmux wrapper script
-curl -LO https://raw.githubusercontent.com/junegunn/fzf/refs/heads/master/bin/fzf-tmux
-chmod +x fzf-tmux
-mv fzf-tmux "$HOME/.local/bin/"
+if [ ! -f "$HOME/.local/bin/fzf-tmux" ]; then
+    curl -LO https://raw.githubusercontent.com/junegunn/fzf/refs/heads/master/bin/fzf-tmux
+    chmod +x fzf-tmux
+    mv fzf-tmux "$HOME/.local/bin/"
+fi
